requirements.txt
Flask==3.1.0
pymongo==4.10.1
python-dotenv==1.0.1
gunicorn==23.0.0
dnspython==2.7.0
mysqlclient==2.2.7
Flask-SQLAlchemy==3.1.1
